server/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
httpx==0.28.0
python-dotenv==1.0.1
groq
pydantic==2.10.3
PyJWT>=2.8.0
cryptography>=42.0.0
pypdf>=5.0.0
pandas>=2.2.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-multipart>=0.0.9
SQLAlchemy>=2.0.36
asyncpg>=0.30.0
aiosqlite>=0.20.0
